Switch to a global mock approach
While simpler to setup I'm really not a fan of Moto 5's new approach
overall. I prefer the approach previous to Moto 5 which was more
explicit and easier to follow overall.

I've run into to various issues making this work. The new approach
uses an event handler that must be established before any clients are.
Sounds simple but the Lambda practice of creating clients in the global
scope can result in issues.

eg. I had to move the dataclass in the function to a separate file
because I needed it for a fixture but that import caused a client to be
established before mocks were enabled.

I've also struggled with relying on chaining fixtures to establish mocks
before clients. I worry about changes to fixtures for whatever reason
resulting in unintended results.